Implementing Json Api on the home page

+++ WeatherApi/GetData.cs
@@ -51,7 +51,7 @@ namespace WeatherApi
                 IList<WeatherData> weatherDataList = jsonArray.Select(p => new WeatherData
                 {
                     dt_txt = (DateTime)p["dt_txt"],
-                    temp = (decimal)p["main"]["temp"],
+                    temp = (int)p["main"]["temp"],
                     icon = (string)p["weather"][0]["icon"]
                 }).ToList();
